refactor: consolidate chat components
