+++ web/src/components/chat/ToolsSelector.tsx
