+++ web/src/components/chat/ui/MinimizeButton.tsx
